+++ src/main/java/com/aoe/restapi/exception/exception/ArgException.java
@@ -1,7 +1,7 @@
 package com.aoe.restapi.exception.exception;
 
 public class ArgException extends RuntimeException {
-    public ArgException(String message) {
-        super(message);
+    public ArgException() {
+        super("argument is not valid");
     }
 }
