+++ src/main/java/com/aoe/restapi/utility/auth/JwtUtil.java
@@ -1,6 +1,5 @@
 package com.aoe.restapi.utility.auth;
 
-import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.security.Keys;
