+++ src/main/java/com/aoe/restapi/utility/http/ResponseUtil.java
@@ -1,4 +1,4 @@
-package com.aoe.restapi.utility;
+package com.aoe.restapi.utility.http;
 
 import com.aoe.restapi.utility.Status.OperationStatus;
 import com.aoe.restapi.utility.Status.OperationStatusSuccess;
